Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

PartiallySeparableNLPModels.Mod_PQN.PartitionedData_TR_PQN

PartiallySeparableNLPModels.Mod_common.Element_function

PartiallySeparableNLPModels.Mod_partitionedNLPModel.PQNNLPModel

PartiallySeparableNLPModels.Mod_partitionedNLPModel.SupportedNLPModel

NLPModels.grad!

NLPModels.obj

PartiallySeparableNLPModels.Mod_PQN.build_PartitionedData_TR_PQN

PartiallySeparableNLPModels.Mod_PQN.update_PQN

PartiallySeparableNLPModels.Mod_PQN.update_PQN!

PartiallySeparableNLPModels.Mod_ab_partitioned_data.evaluate_grad_part_data

PartiallySeparableNLPModels.Mod_ab_partitioned_data.evaluate_grad_part_data!

PartiallySeparableNLPModels.Mod_ab_partitioned_data.evaluate_obj_part_data

PartiallySeparableNLPModels.Mod_ab_partitioned_data.evaluate_obj_part_data!

PartiallySeparableNLPModels.Mod_ab_partitioned_data.evaluate_y_part_data!

PartiallySeparableNLPModels.Mod_ab_partitioned_data.product_part_data_x

PartiallySeparableNLPModels.Mod_ab_partitioned_data.product_part_data_x!

PartiallySeparableNLPModels.Mod_ab_partitioned_data.update_nlp!

PartiallySeparableNLPModels.Mod_common.compiled_grad_element_function

PartiallySeparableNLPModels.Mod_common.distinct_element_expr_tree

PartiallySeparableNLPModels.Mod_partitionedNLPModel.get_expr_tree

​

PartiallySeparableNLPModels.Mod_ab_partitioned_data.evaluate_grad_part_data! — Method

evaluate_grad_part_data!(g::Vector{Y}, part_data::PartitionedData, x::Vector{Y}) where {Y <: Number}
evaluate_grad_part_data!(part_data::PartitionedData)

Evaluate in place of g the gradient at the point x after the computation of the element gradients (stored in part_data.pg). When g and x are omitted, it considers that part_data.g and part_data.x are respectively g and x.

PartiallySeparableNLPModels.Mod_ab_partitioned_data.evaluate_grad_part_data — Method

gradient = evaluate_grad_part_data(part_data::PartitionedData, x::Vector{Y}) where {Y <: Number}

Return the gradient vector g at the point x after the computation of the element gradients (stored in part_data.pg).

PartiallySeparableNLPModels.Mod_ab_partitioned_data.evaluate_obj_part_data! — Method

evaluate_obj_part_data!(part_data::PartitionedData)

Compute and store the partially separable objective value stored in part_data, as a sum of element functions, in part_data given part_data.x.

PartiallySeparableNLPModels.Mod_ab_partitioned_data.evaluate_obj_part_data — Method

fx = evaluate_obj_part_data(part_data::PartitionedData, x::Vector{Y}) where {Y <: Number}

Return the partially separable objective value stored in part_data, computed as a sum of element functions.

PartiallySeparableNLPModels.Mod_ab_partitioned_data.evaluate_y_part_data! — Method

evaluate_y_part_data!(part_data::PartitionedData, x::Vector{Y}, s::Vector{Y}) where {Y <: Number}
evaluate_y_part_data!(part_data::PartitionedData, s::Vector{Y}) where {Y <: Number}

Compute the element gradients differences such as ∇̂fᵢ(x+s)-∇̂fᵢ(x) for each element functions. It stores the results in part_data.pv. evaluate_y_part_data!(part_data, s) consider the point part_data.x and the partitioned gradient part_data.pg known.

PartiallySeparableNLPModels.Mod_ab_partitioned_data.product_part_data_x! — Method

product_part_data_x!(res::Vector{Y}, part_data::PartitionedData, x::Vector{Y}) where {Y <: Number}
product_part_data_x!(epv_res::PartitionedStructures.Elemental_pv{Y}, part_data::PartitionedData, epv::PartitionedStructures.Elemental_pv{Y}) where {Y <: Number}
product_part_data_x!(epv_res::PartitionedStructures.Elemental_pv{Y}, pB::T, epv::PartitionedStructures.Elemental_pv{Y}) where {Y <: Number, T <: PartitionedStructures.Part_mat{Y}}

Return in place of res the product between the partitioned matrix part_data.pB and x.

PartiallySeparableNLPModels.Mod_ab_partitioned_data.product_part_data_x — Method

Bx = product_part_data_x(part_data::PartitionedData, x::Vector{Y}) where {Y <: Number}

Return the product between the partitioned matrix part_data.pB and x.

PartiallySeparableNLPModels.Mod_common.Element_function — Type

Element_function

Gather the informations indentifying an element function in a PQNNLPModel, and its particular properties. Element_function has fields:

i describing the index of the element function;

index_element_tree the index occupied in the element-function vector after the deletion of redundant element functions;

variable_indices informing the elemental variable of Element_function

type describing Element_function as constant, linear, quadratic, cubic or more non linear;

convexity_status describing Element_function as constant, linear, convex, concave or unknown.

PartiallySeparableNLPModels.Mod_common.compiled_grad_element_function — Method

element_gradient_tape = compiled_grad_element_function(element_function::T; ni::Int = length(ExpressionTreeForge.get_elemental_variables(element_function)), type = Float64) where {T}

Return the elment_gradient_tape::GradientTape which speed up the gradient computation of element_function with ReverseDiff.

PartiallySeparableNLPModels.Mod_common.distinct_element_expr_tree — Method

(element_expr_trees, indices_element_tree) = distinct_element_expr_tree(vec_element_expr_tree::Vector{T}, vec_element_variables::Vector{Vector{Int}}; N::Int = length(vec_element_expr_tree)) where {T}

In practice, there may have several element functions having the same expression tree. distinct_element_expr_tree filters the vector vec_element_expr_tree to return element_expr_trees the distincts element functions. length(element_expr_trees) == M < N == length(vec_element_expr_tree). In addition it returns indices_element_tree, who records the index (1 <= i <= M) related ot the expression tree of each element function.

PartiallySeparableNLPModels.Mod_partitionedNLPModel.SupportedNLPModel — Type

Accumulate the supported NLPModels.

PartiallySeparableNLPModels.Mod_partitionedNLPModel.PQNNLPModel — Type

PQNNLPModel{ T, S, G, M <: AbstractNLPModel{T, S}, Meta <: AbstractNLPModelMeta{T, S}} <: AbstractPQNNLPModel{T, S}

A partitioned quasi-Newton NLPModel. A PQNNLPModel has field:

meta counting numerous information about the PQNNLPModel;

part_data allocate the partitioned structures required by a partitioned quasi-Newton trust-region method;

nlp the original NLPModel.

NLPModels.grad! — Method

g = grad!(nlp, x, g)

Evaluate ∇f(x), the gradient of the objective function at x in place.

NLPModels.obj — Method

f = obj(nlp, x)

Evaluate f(x), the objective function of nlp at x.

PartiallySeparableNLPModels.Mod_partitionedNLPModel.get_expr_tree — Method

expr, n, x0 = get_expr_tree(adnlp::MathOptNLPModel; x0::Vector{T} = copy(adnlp.meta.x0), kwargs...) where {T <: Number}
expr, n, x0 = get_expr_tree(adnlp::ADNLPModel; x0::Vector{T} = copy(adnlp.meta.x0), kwargs...) where {T <: Number}

Return for a MathOptNLPModel or a ADNLPModel: the expression tree of the objective function expr::Expr, the size of the problem n and the initial point x0.

PartiallySeparableNLPModels.Mod_PQN.PartitionedData_TR_PQN — Type

PartitionedData_TR_PQN{G, T <: Number, P <: Part_mat{T}} <: PartitionedData

Gather the structures required to run a partitioned quasi-Newton trust-region method. PartitionedData_TR_PQN has fields:

n the size of the problem;

N the number of element functions;

vec_elt_fun a Element_function vector, of size N;

M the number of distinct element-function expression trees;

vec_elt_complete_expr_tree a Complete_expr_tree vector, of size M;

element_expr_tree_table a vector of size M, the i-th element element_expr_tree_table[i]::Vector{Int} informs which element functions have the vec_elt_complete_expr_tree[i] expreesion tree;

index_element_tree a vector of size N where each component indicates which Complete_expr_tree from vec_elt_complete_expr_tree use for the corresponding element;

vec_compiled_element_gradients the vector gathering the compiled tapes for the element gradient evaluations;

x the current point;

v a temporary vector;

s the current step;

pg the partitioned gradient;

pv a temporary partitioned vector;

py the partitioned gradient difference;

ps the partitioned step;

pB the partitioned matrix (main memory cost);

fx the current value of the objective function;

name the name of partitioned quasi-Newton update peformed at each iterate.

PartiallySeparableNLPModels.Mod_PQN.build_PartitionedData_TR_PQN — Method

partitioneddata_tr_pqn = build_PartitionedData_TR_PQN(expr_tree, n)

Return the structure required to run a partitioned quasi-Newton trust-region method. It finds the partially-separable structure of an expression tree expr_tree representing f(x) = ∑fᵢ(xᵢ). Then it allocates the partitioned structures required. Note that to define properly the sparse matrix of the partitioned matrix we need the size of the problem: n.

PartiallySeparableNLPModels.Mod_PQN.update_PQN! — Method

update_PQN!(pd_pqn::PartitionedData_TR_PQN{G, T, P}, s::Vector{T})
update_PQN!(pd_pqn::PartitionedData_TR_PQN{G, T, P}, x::Vector{T}, s::Vector{T})

Perform the partitioned quasi-Newton update given the vectors x and s. When x is omitted, update_PQN! consider that pd_pqn already know the value of x. Moreover, it assumes that the partitioned gradient at x is already computed in pd_pqn.pg.

PartiallySeparableNLPModels.Mod_PQN.update_PQN — Method

B = update_PQN(  pd_pqn::PartitionedData_TR_PQN{G, T, P}, x::Vector{T}, s::Vector{T};

Perform the partitioned quasi-Newton update given the vectors x and s.

PartiallySeparableNLPModels.Mod_ab_partitioned_data.update_nlp! — Method

update_nlp!(pd_pqn::PartitionedData_TR_PQN{G, T, P}, s::Vector{T})
update_nlp!(pd_pqn::PartitionedData_TR_PQN{G, T, P}, x::Vector{T}, s::Vector{T})

Perform the partitioned quasi-Newton update given the vectors x and s. When x is omitted, update_PQN! consider that pd_pqn already know the value of x. Moreover, it assumes that the partitioned gradient at x is already computed in pd_pqn.pg.