Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

PartiallySeparableNLPModels.ModAbstractPSNLPModels.ElementFunction

PartiallySeparableNLPModels.ModAbstractPSNLPModels.SupportedNLPModel

PartiallySeparableNLPModels.ModPBFGSNLPModels.PBFGSNLPModel

PartiallySeparableNLPModels.ModPCSNLPModels.PCSNLPModel

PartiallySeparableNLPModels.ModPLBFGSNLPModels.PLBFGSNLPModel

PartiallySeparableNLPModels.ModPLSENLPModels.PLSENLPModel

PartiallySeparableNLPModels.ModPLSR1NLPModels.PLSR1NLPModel

PartiallySeparableNLPModels.ModPSENLPModels.PSENLPModel

PartiallySeparableNLPModels.ModPSNLPModels.PSNLPModel

PartiallySeparableNLPModels.ModPSR1NLPModels.PSR1NLPModel

NLPModels.grad!

NLPModels.hprod!

NLPModels.hprod!

NLPModels.obj

PartiallySeparableNLPModels.ModAbstractPSNLPModels.evaluate_grad_part_data

PartiallySeparableNLPModels.ModAbstractPSNLPModels.evaluate_grad_part_data!

PartiallySeparableNLPModels.ModAbstractPSNLPModels.evaluate_obj_part_data

PartiallySeparableNLPModels.ModAbstractPSNLPModels.evaluate_obj_part_data!

PartiallySeparableNLPModels.ModAbstractPSNLPModels.evaluate_y_part_data!

PartiallySeparableNLPModels.ModAbstractPSNLPModels.hess_approx

PartiallySeparableNLPModels.ModAbstractPSNLPModels.partitionedMulOp!

PartiallySeparableNLPModels.ModAbstractPSNLPModels.product_part_data_x

PartiallySeparableNLPModels.ModAbstractPSNLPModels.product_part_data_x!

PartiallySeparableNLPModels.ModAbstractPSNLPModels.update_nlp

PartiallySeparableNLPModels.ModAbstractPSNLPModels.update_nlp!

PartiallySeparableNLPModels.Utils.compiled_grad_element_function

PartiallySeparableNLPModels.Utils.distinct_element_expr_tree

PartiallySeparableNLPModels.Utils.partially_separable_structure

​

PartiallySeparableNLPModels.ModAbstractPSNLPModels.SupportedNLPModel — Type

Accumulate the supported NLPModels.

PartiallySeparableNLPModels.ModAbstractPSNLPModels.ElementFunction — Type

ElementFunction

A type that gathers the information indentifying an element function in a PartiallySeparableNLPModel, and its properties. ElementFunction has fields:

i: the index of the element function;

index_element_tree: the index occupied in the element-function vector after the deletion of redundant element functions;

variable_indices: list of elemental variables of ElementFunction;

type: constant, linear, quadratic, cubic or general;

convexity_status: constant, linear, convex, concave or unknown.

NLPModels.grad! — Method

g = grad!(nlp, x, g)

Evaluate ∇f(x), the gradient of the objective function at x in place.

NLPModels.hprod! — Method

hprod!(nlp::AbstractPQNNLPModel, x::AbstractVector, v::AbstractVector, Hv::AbstractVector; obj_weight=1.)

Evaluate the product of the objective Hessian at x with the vector v, with objective function scaled by obj_weight.

NLPModels.hprod! — Method

hprod!(nlp::AbstractPartiallySeparableNLPModel, x::AbstractVector, v::AbstractVector, Hv::AbstractVector; obj_weight=1.)

Evaluate the product of the objective Hessian at x with the vector v, with objective function scaled by obj_weight.

NLPModels.obj — Method

f = obj(nlp, x)

Evaluate f(x), the objective function of nlp at x.

PartiallySeparableNLPModels.ModAbstractPSNLPModels.evaluate_grad_part_data! — Method

evaluate_grad_part_data!(g::AbstractVector{Y}, psnlp::AbstractPartiallySeparableNLPModel, x::AbstractVector{Y}) where {Y <: Number}
evaluate_grad_part_data!(psnlp::AbstractPartiallySeparableNLPModel)

Evaluate the gradient at x in place. It computes and store the element gradients in psnlp.pg and accumulate their contributions in g. When g and x are omitted, consider that psnlp.g and psnlp.x are g and x, respectively.

PartiallySeparableNLPModels.ModAbstractPSNLPModels.evaluate_grad_part_data — Method

gradient = evaluate_grad_part_data(psnlp::AbstractPartiallySeparableNLPModel, x::AbstractVector{Y}) where {Y <: Number}

Return the gradient vector g at x. After the computation of the element gradients (stored in psnlp.pg), g is built by accumulating their contributions.

PartiallySeparableNLPModels.ModAbstractPSNLPModels.evaluate_obj_part_data! — Method

evaluate_obj_part_data!(psnlp::AbstractPartiallySeparableNLPModel)

Compute the partially separable objective function, as a sum of element functions, at psnlp.x. Its objective value is stored in psnlp.fx .

PartiallySeparableNLPModels.ModAbstractPSNLPModels.evaluate_obj_part_data — Method

fx = evaluate_obj_part_data(psnlp::AbstractPartiallySeparableNLPModel, x::AbstractVector{Y}) where {Y <: Number}

Return the partially separable objective at x.

PartiallySeparableNLPModels.ModAbstractPSNLPModels.evaluate_y_part_data! — Method

evaluate_y_part_data!(psnlp::AbstractPartiallySeparableNLPModel, x::AbstractVector{Y}, s::AbstractVector{Y}) where {Y <: Number}
evaluate_y_part_data!(psnlp::AbstractPartiallySeparableNLPModel, s::AbstractVector{Y}) where {Y <: Number}

Compute element gradients differences ∇̂fᵢ(x+s)-∇̂fᵢ(x) for each element function. Store the results in psnlp.py. When x is ommited, consider psnlp.x as x and psnlp.pg as the partitioned gradient at x.

PartiallySeparableNLPModels.ModAbstractPSNLPModels.hess_approx — Method

B = hess_approx(nlp::AbstractPQNNLPModel)

Return the Hessian approximation of nlp.

PartiallySeparableNLPModels.ModAbstractPSNLPModels.partitionedMulOp! — Method

partitionedMulOp!(pqn_nlp::AbstractPQNNLPModel, res, v, α, β)

Partitioned 5-arg mul! for pqn_nlp using the partitioned matrix and partitioned vectors to destribute and collect the result of element matrix-vector products.

PartiallySeparableNLPModels.ModAbstractPSNLPModels.product_part_data_x! — Method

product_part_data_x!(res::AbstractVector{Y}, psnlp::AbstractPartiallySeparableNLPModel, v::AbstractVector{Y}) where {Y <: Number}
product_part_data_x!(epv_res::PartitionedStructures.Elemental_pv{Y}, psnlp::AbstractPartiallySeparableNLPModel, epv::PartitionedStructures.Elemental_pv{Y}) where {Y <: Number}
product_part_data_x!(epv_res::PartitionedStructures.Elemental_pv{Y}, pB::T, epv::PartitionedStructures.Elemental_pv{Y}) where {Y <: Number, T <: PartitionedStructures.Part_mat{Y}}

Store the product between the partitioned matrix psnlp.pB and the vector v in res. The computation of every element matrix-vector product require two partitioned vectors. The first partitioned vector epv is created from v and the results are stored in the second partitioned vector epv_res which builds the value store by g.

PartiallySeparableNLPModels.ModAbstractPSNLPModels.product_part_data_x — Method

Bx = product_part_data_x(psnlp::AbstractPartiallySeparableNLPModel, v::AbstractVector{Y}) where {Y <: Number}

Return the product between the partitioned matrix psnlp.pB and v.

PartiallySeparableNLPModels.ModAbstractPSNLPModels.update_nlp! — Method

update_nlp!(pqn_nlp::AbstractPQNNLPModel{T,S}, s::Vector{T})
update_nlp!(pqn_nlp::AbstractPQNNLPModel{T,S}, x::Vector{T}, s::Vector{T})

Perform the partitioned quasi-Newton update given the current point x and the step s. When x is omitted, update_PQN! consider that pqn_nlp has the current point in pqnnlp.x. Moreover, it assumes that the partitioned gradient atxis already computed inpqnnlp.pg. Will be replace bypush!whenPartitionedVector` are implemented.

PartiallySeparableNLPModels.ModAbstractPSNLPModels.update_nlp — Method

update_nlp(pqn_nlp::AbstractPQNNLPModel{T,S}, x::Vector{T}, s::Vector{T})

Perform the partitioned quasi-Newton update given the current point x and the step s. When x is omitted, update_PQN! consider that pqn_nlp has the current point in pqnnlp.x. Moreover, it assumes that the partitioned gradient atxis already computed inpqnnlp.pg. Will be replace bypush!whenPartitionedVector` are implemented.

PartiallySeparableNLPModels.Utils.compiled_grad_element_function — Method

element_gradient_tape = compiled_grad_element_function(element_function::T; ni::Int = length(ExpressionTreeForge.get_elemental_variables(element_function)), type = Float64) where {T}

Return the elment_gradient_tape::GradientTape which speed up the gradient computation of element_function with ReverseDiff.

PartiallySeparableNLPModels.Utils.distinct_element_expr_tree — Method

(element_expr_trees, indices_element_tree) = distinct_element_expr_tree(vec_element_expr_tree::Vector{T}, vec_element_variables::Vector{Vector{Int}}; N::Int = length(vec_element_expr_tree)) where {T}

In practice, there may have several element functions having the same expression tree. distinct_element_expr_tree filters the vector vec_element_expr_tree to return element_expr_trees the distincts element functions. length(element_expr_trees) == M < N == length(vec_element_expr_tree). In addition it returns indices_element_tree, who records the index (1 <= i <= M) related ot the expression tree of each element function.

PartiallySeparableNLPModels.Utils.partially_separable_structure — Method

partitioneddata_tr_pqn = build_PartitionedDataTRPQN(expr_tree, n)

Return the structure required to run a partitioned quasi-Newton trust-region method. It finds the partially-separable structure of an expression tree expr_tree representing f(x) = ∑fᵢ(xᵢ). Then it allocates the partitioned structures required. To define properly the sparse matrix of the partitioned matrix we need the size of the problem: n.

PartiallySeparableNLPModels.ModPBFGSNLPModels.PBFGSNLPModel — Type

PBFGSNLPModel{G, P, T, S, M <: AbstractNLPModel{T, S}, Meta <: AbstractNLPModelMeta{T, S},} <: AbstractPQNNLPModel{T,S}

Deduct and allocate the partitioned structures of a NLPModel using partitioned BFGS Hessian approximation. PBFGSNLPModel has fields:

nlp: the original model;

meta: gather information about the PartiallySeparableNLPModel;

counters: count how many standards methods of NLPModels are called;

n: the size of the problem;

N: the number of element functions;

vec_elt_fun: a ElementFunction vector, of size N;

M: the number of distinct element-function expression trees;

vec_elt_complete_expr_tree: a Complete_expr_tree vector, of size M;

element_expr_tree_table: a vector of size M, the i-th element element_expr_tree_table[i]::Vector{Int} informs which element functions use the vec_elt_complete_expr_tree[i] expression tree;

index_element_tree: a vector of size N where each component indicates which Complete_expr_tree from vec_elt_complete_expr_tree is used for the corresponding element;

vec_compiled_element_gradients: the vector gathering the compiled tapes for every element gradient evaluations;

x: the current point;

v: a temporary vector;

s: the current step;

pg: the partitioned gradient;

pv: a temporary partitioned vector;

py: the partitioned gradient difference;

ps: the partitioned step;

phv: the partitioned Hessian-vector product;

pB: the partitioned matrix (main memory cost);

fx: the current value of the objective function;

name: the name of partitioned quasi-Newton update performed

PartiallySeparableNLPModels.ModPCSNLPModels.PCSNLPModel — Type

PCSNLPModel{G, P, T, S, M <: AbstractNLPModel{T, S}, Meta <: AbstractNLPModelMeta{T, S},} <: AbstractPQNNLPModel{T,S}

Deduct and allocate the partitioned structures of a NLPModel using partitioned BFGS Hessian approximation. PCSNLPModel has fields:

nlp: the original model;

meta: gather information about the PartiallySeparableNLPModel;

counters: count how many standards methods of NLPModels are called;

n: the size of the problem;

N: the number of element functions;

vec_elt_fun: a ElementFunction vector, of size N;

M: the number of distinct element-function expression trees;

vec_elt_complete_expr_tree: a Complete_expr_tree vector, of size M;

element_expr_tree_table: a vector of size M, the i-th element element_expr_tree_table[i]::Vector{Int} informs which element functions use the vec_elt_complete_expr_tree[i] expression tree;

index_element_tree: a vector of size N where each component indicates which Complete_expr_tree from vec_elt_complete_expr_tree is used for the corresponding element;

vec_compiled_element_gradients: the vector gathering the compiled tapes for every element gradient evaluations;

x: the current point;

v: a temporary vector;

s: the current step;

pg: the partitioned gradient;

pv: a temporary partitioned vector;

py: the partitioned gradient difference;

ps: the partitioned step;

phv: the partitioned Hessian-vector product;

pB: the partitioned matrix (main memory cost);

fx: the current value of the objective function;

name: the name of partitioned quasi-Newton update performed

PartiallySeparableNLPModels.ModPLBFGSNLPModels.PLBFGSNLPModel — Type

PLBFGSNLPModel{G, P, T, S, M <: AbstractNLPModel{T, S}, Meta <: AbstractNLPModelMeta{T, S},} <: AbstractPQNNLPModel{T,S}

Deduct and allocate the partitioned structures of a NLPModel using partitioned BFGS Hessian approximation. PLBFGSNLPModel has fields:

nlp: the original model;

meta: gather information about the PartiallySeparableNLPModel;

counters: count how many standards methods of NLPModels are called;

n: the size of the problem;

N: the number of element functions;

vec_elt_fun: a ElementFunction vector, of size N;

M: the number of distinct element-function expression trees;

vec_elt_complete_expr_tree: a Complete_expr_tree vector, of size M;

element_expr_tree_table: a vector of size M, the i-th element element_expr_tree_table[i]::Vector{Int} informs which element functions use the vec_elt_complete_expr_tree[i] expression tree;

index_element_tree: a vector of size N where each component indicates which Complete_expr_tree from vec_elt_complete_expr_tree is used for the corresponding element;

vec_compiled_element_gradients: the vector gathering the compiled tapes for every element gradient evaluations;

x: the current point;

v: a temporary vector;

s: the current step;

pg: the partitioned gradient;

pv: a temporary partitioned vector;

py: the partitioned gradient difference;

ps: the partitioned step;

phv: the partitioned Hessian-vector product;

pB: the partitioned matrix (main memory cost);

fx: the current value of the objective function;

name: the name of partitioned quasi-Newton update performed

PartiallySeparableNLPModels.ModPLSENLPModels.PLSENLPModel — Type

PLSENLPModel{G, P, T, S, M <: AbstractNLPModel{T, S}, Meta <: AbstractNLPModelMeta{T, S},} <: AbstractPQNNLPModel{T,S}

Deduct and allocate the partitioned structures of a NLPModel using partitioned BFGS Hessian approximation. PLSENLPModel has fields:

nlp: the original model;

meta: gather information about the PartiallySeparableNLPModel;

counters: count how many standards methods of NLPModels are called;

n: the size of the problem;

N: the number of element functions;

vec_elt_fun: a ElementFunction vector, of size N;

M: the number of distinct element-function expression trees;

vec_elt_complete_expr_tree: a Complete_expr_tree vector, of size M;

element_expr_tree_table: a vector of size M, the i-th element element_expr_tree_table[i]::Vector{Int} informs which element functions use the vec_elt_complete_expr_tree[i] expression tree;

index_element_tree: a vector of size N where each component indicates which Complete_expr_tree from vec_elt_complete_expr_tree is used for the corresponding element;

vec_compiled_element_gradients: the vector gathering the compiled tapes for every element gradient evaluations;

x: the current point;

v: a temporary vector;

s: the current step;

pg: the partitioned gradient;

pv: a temporary partitioned vector;

py: the partitioned gradient difference;

ps: the partitioned step;

phv: the partitioned Hessian-vector product;

pB: the partitioned matrix (main memory cost);

fx: the current value of the objective function;

name: the name of partitioned quasi-Newton update performed

PartiallySeparableNLPModels.ModPLSR1NLPModels.PLSR1NLPModel — Type

PLSR1NLPModel{G, P, T, S, M <: AbstractNLPModel{T, S}, Meta <: AbstractNLPModelMeta{T, S},} <: AbstractPQNNLPModel{T,S}

Deduct and allocate the partitioned structures of a NLPModel using partitioned BFGS Hessian approximation. PLSR1NLPModel has fields:

nlp: the original model;

meta: gather information about the PartiallySeparableNLPModel;

counters: count how many standards methods of NLPModels are called;

n: the size of the problem;

N: the number of element functions;

vec_elt_fun: a ElementFunction vector, of size N;

M: the number of distinct element-function expression trees;

vec_elt_complete_expr_tree: a Complete_expr_tree vector, of size M;

element_expr_tree_table: a vector of size M, the i-th element element_expr_tree_table[i]::Vector{Int} informs which element functions use the vec_elt_complete_expr_tree[i] expression tree;

index_element_tree: a vector of size N where each component indicates which Complete_expr_tree from vec_elt_complete_expr_tree is used for the corresponding element;

vec_compiled_element_gradients: the vector gathering the compiled tapes for every element gradient evaluations;

x: the current point;

v: a temporary vector;

s: the current step;

pg: the partitioned gradient;

pv: a temporary partitioned vector;

py: the partitioned gradient difference;

ps: the partitioned step;

phv: the partitioned Hessian-vector product;

pB: the partitioned matrix (main memory cost);

fx: the current value of the objective function;

name: the name of partitioned quasi-Newton update performed

PartiallySeparableNLPModels.ModPSENLPModels.PSENLPModel — Type

PSENLPModel{G, P, T, S, M <: AbstractNLPModel{T, S}, Meta <: AbstractNLPModelMeta{T, S},} <: AbstractPQNNLPModel{T,S}

Deduct and allocate the partitioned structures of a NLPModel using partitioned BFGS Hessian approximation. PSENLPModel has fields:

nlp: the original model;

meta: gather information about the PartiallySeparableNLPModel;

counters: count how many standards methods of NLPModels are called;

n: the size of the problem;

N: the number of element functions;

vec_elt_fun: a ElementFunction vector, of size N;

M: the number of distinct element-function expression trees;

vec_elt_complete_expr_tree: a Complete_expr_tree vector, of size M;

element_expr_tree_table: a vector of size M, the i-th element element_expr_tree_table[i]::Vector{Int} informs which element functions use the vec_elt_complete_expr_tree[i] expression tree;

index_element_tree: a vector of size N where each component indicates which Complete_expr_tree from vec_elt_complete_expr_tree is used for the corresponding element;

vec_compiled_element_gradients: the vector gathering the compiled tapes for every element gradient evaluations;

x: the current point;

v: a temporary vector;

s: the current step;

pg: the partitioned gradient;

pv: a temporary partitioned vector;

py: the partitioned gradient difference;

ps: the partitioned step;

phv: the partitioned Hessian-vector product;

pB: the partitioned matrix (main memory cost);

fx: the current value of the objective function;

name: the name of partitioned quasi-Newton update performed

PartiallySeparableNLPModels.ModPSNLPModels.PSNLPModel — Type

PSNLPModel{G, T, S, M <: AbstractNLPModel{T, S}, Meta <: AbstractNLPModelMeta{T, S},} <: AbstractPQNNLPModel{T,S}

Deduct and allocate the partitioned structures of a NLPModel using partitioned BFGS Hessian approximation. PSNLPModel has fields:

nlp: the original model;

meta: gather information about the PartiallySeparableNLPModel;

counters: count how many standards methods of NLPModels are called;

n: the size of the problem;

N: the number of element functions;

vec_elt_fun: a ElementFunction vector, of size N;

M: the number of distinct element-function expression trees;

vec_elt_complete_expr_tree: a Complete_expr_tree vector, of size M;

element_expr_tree_table: a vector of size M, the i-th element element_expr_tree_table[i]::Vector{Int} informs which element functions use the vec_elt_complete_expr_tree[i] expression tree;

index_element_tree: a vector of size N where each component indicates which Complete_expr_tree from vec_elt_complete_expr_tree is used for the corresponding element;

vec_compiled_element_gradients: the vector gathering the compiled tapes for every element gradient evaluations;

x: the current point;

v: a temporary vector;

s: the current step;

pg: the partitioned gradient;

pv: a temporary partitioned vector;

py: the partitioned gradient difference;

ps: the partitioned step;

phv: the partitioned Hessian-vector product;

fx: the current value of the objective function;

name: the name of partitioned quasi-Newton update performed

PartiallySeparableNLPModels.ModPSR1NLPModels.PSR1NLPModel — Type

PSR1NLPModel{G, P, T, S, M <: AbstractNLPModel{T, S}, Meta <: AbstractNLPModelMeta{T, S},} <: AbstractPQNNLPModel{T,S}

Deduct and allocate the partitioned structures of a NLPModel using partitioned BFGS Hessian approximation. PSR1NLPModel has fields:

nlp: the original model;

meta: gather information about the PartiallySeparableNLPModel;

counters: count how many standards methods of NLPModels are called;

n: the size of the problem;

N: the number of element functions;

vec_elt_fun: a ElementFunction vector, of size N;

M: the number of distinct element-function expression trees;

vec_elt_complete_expr_tree: a Complete_expr_tree vector, of size M;

element_expr_tree_table: a vector of size M, the i-th element element_expr_tree_table[i]::Vector{Int} informs which element functions use the vec_elt_complete_expr_tree[i] expression tree;

index_element_tree: a vector of size N where each component indicates which Complete_expr_tree from vec_elt_complete_expr_tree is used for the corresponding element;

vec_compiled_element_gradients: the vector gathering the compiled tapes for every element gradient evaluations;

x: the current point;

v: a temporary vector;

s: the current step;

pg: the partitioned gradient;

pv: a temporary partitioned vector;

py: the partitioned gradient difference;

ps: the partitioned step;

phv: the partitioned Hessian-vector product;

pB: the partitioned matrix (main memory cost);

fx: the current value of the objective function;

name: the name of partitioned quasi-Newton update performed