Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

PartiallySeparableNLPModels.Mod_PQN.PartitionedDataTRPQN

PartiallySeparableNLPModels.Mod_common.ElementFunction

PartiallySeparableNLPModels.Mod_partitionedNLPModel.PartiallySeparableNLPModel

PartiallySeparableNLPModels.Mod_partitionedNLPModel.SupportedNLPModel

NLPModels.grad!

NLPModels.hprod

NLPModels.hprod!

NLPModels.hprod!

NLPModels.obj

PartiallySeparableNLPModels.Mod_PQN.build_PartitionedDataTRPQN

PartiallySeparableNLPModels.Mod_PQN.update_PQN

PartiallySeparableNLPModels.Mod_PQN.update_PQN!

PartiallySeparableNLPModels.Mod_ab_partitioned_data.evaluate_grad_part_data

PartiallySeparableNLPModels.Mod_ab_partitioned_data.evaluate_grad_part_data!

PartiallySeparableNLPModels.Mod_ab_partitioned_data.evaluate_obj_part_data

PartiallySeparableNLPModels.Mod_ab_partitioned_data.evaluate_obj_part_data!

PartiallySeparableNLPModels.Mod_ab_partitioned_data.evaluate_y_part_data!

PartiallySeparableNLPModels.Mod_ab_partitioned_data.product_part_data_x

PartiallySeparableNLPModels.Mod_ab_partitioned_data.product_part_data_x!

PartiallySeparableNLPModels.Mod_ab_partitioned_data.update_nlp!

PartiallySeparableNLPModels.Mod_common.compiled_grad_element_function

PartiallySeparableNLPModels.Mod_common.distinct_element_expr_tree

PartiallySeparableNLPModels.Mod_partitionedNLPModel.get_expr_tree

PartiallySeparableNLPModels.Mod_partitionedNLPModel.hess_approx

​

NLPModels.hprod! — Method

hprod!(part_data::PartitionedData, x::AbstractVector, v::AbstractVector, hv::AbstractVector)

Set hv::Vector to the product between the partitioned hessian ∇²f(x) and the vector v.

NLPModels.hprod — Method

hv = hprod(part_data::PartitionedData, x::AbstractVector, v::AbstractVector)

Return hv::Vector the product between the partitioned hessian ∇²f(x) and the vector v.

PartiallySeparableNLPModels.Mod_ab_partitioned_data.evaluate_grad_part_data! — Method

evaluate_grad_part_data!(g::AbstractVector{Y}, part_data::PartitionedData, x::AbstractVector{Y}) where {Y <: Number}
evaluate_grad_part_data!(part_data::PartitionedData)

Evaluate the gradient at x in place. It computes and store the element gradients in part_data.pg and accumulate their contributions in g. When g and x are omitted, consider that part_data.g and part_data.x are g and x, respectively.

PartiallySeparableNLPModels.Mod_ab_partitioned_data.evaluate_grad_part_data — Method

gradient = evaluate_grad_part_data(part_data::PartitionedData, x::AbstractVector{Y}) where {Y <: Number}

Return the gradient vector g at x. After the computation of the element gradients (stored in part_data.pg), g is built by accumulating their contributions.

PartiallySeparableNLPModels.Mod_ab_partitioned_data.evaluate_obj_part_data! — Method

evaluate_obj_part_data!(part_data::PartitionedData)

Compute the partially separable objective function, as a sum of element functions, at part_data.x. Its objective value is stored in part_data.fx .

PartiallySeparableNLPModels.Mod_ab_partitioned_data.evaluate_obj_part_data — Method

fx = evaluate_obj_part_data(part_data::PartitionedData, x::AbstractVector{Y}) where {Y <: Number}

Return the partially separable objective at x.

PartiallySeparableNLPModels.Mod_ab_partitioned_data.evaluate_y_part_data! — Method

evaluate_y_part_data!(part_data::PartitionedData, x::AbstractVector{Y}, s::AbstractVector{Y}) where {Y <: Number}
evaluate_y_part_data!(part_data::PartitionedData, s::AbstractVector{Y}) where {Y <: Number}

Compute element gradients differences ∇̂fᵢ(x+s)-∇̂fᵢ(x) for each element function. Store the results in part_data.py. When x is ommited, consider part_data.x as x and part_data.pg as the partitioned gradient at x.

PartiallySeparableNLPModels.Mod_ab_partitioned_data.product_part_data_x! — Method

product_part_data_x!(res::AbstractVector{Y}, part_data::PartitionedData, v::AbstractVector{Y}) where {Y <: Number}
product_part_data_x!(epv_res::PartitionedStructures.Elemental_pv{Y}, part_data::PartitionedData, epv::PartitionedStructures.Elemental_pv{Y}) where {Y <: Number}
product_part_data_x!(epv_res::PartitionedStructures.Elemental_pv{Y}, pB::T, epv::PartitionedStructures.Elemental_pv{Y}) where {Y <: Number, T <: PartitionedStructures.Part_mat{Y}}

Store the product between the partitioned matrix part_data.pB and the vector v in res. The computation of every element matrix-vector product require two partitioned vectors. The first partitioned vector epv is created from v and the results are stored in the second partitioned vector epv_res which builds the value store by g.

PartiallySeparableNLPModels.Mod_ab_partitioned_data.product_part_data_x — Method

Bx = product_part_data_x(part_data::PartitionedData, v::AbstractVector{Y}) where {Y <: Number}

Return the product between the partitioned matrix part_data.pB and v.

PartiallySeparableNLPModels.Mod_common.ElementFunction — Type

ElementFunction

A type that gathers the information indentifying an element function in a PartiallySeparableNLPModel, and its properties. ElementFunction has fields:

i: the index of the element function;

index_element_tree: the index occupied in the element-function vector after the deletion of redundant element functions;

variable_indices: list of elemental variables of ElementFunction;

type: constant, linear, quadratic, cubic or general;

convexity_status: constant, linear, convex, concave or unknown.

PartiallySeparableNLPModels.Mod_common.compiled_grad_element_function — Method

element_gradient_tape = compiled_grad_element_function(element_function::T; ni::Int = length(ExpressionTreeForge.get_elemental_variables(element_function)), type = Float64) where {T}

Return the elment_gradient_tape::GradientTape which speed up the gradient computation of element_function with ReverseDiff.

PartiallySeparableNLPModels.Mod_common.distinct_element_expr_tree — Method

(element_expr_trees, indices_element_tree) = distinct_element_expr_tree(vec_element_expr_tree::Vector{T}, vec_element_variables::Vector{Vector{Int}}; N::Int = length(vec_element_expr_tree)) where {T}

In practice, there may have several element functions having the same expression tree. distinct_element_expr_tree filters the vector vec_element_expr_tree to return element_expr_trees the distincts element functions. length(element_expr_trees) == M < N == length(vec_element_expr_tree). In addition it returns indices_element_tree, who records the index (1 <= i <= M) related ot the expression tree of each element function.

PartiallySeparableNLPModels.Mod_partitionedNLPModel.SupportedNLPModel — Type

Accumulate the supported NLPModels.

PartiallySeparableNLPModels.Mod_partitionedNLPModel.PartiallySeparableNLPModel — Type

PartiallySeparableNLPModel{ T, S, G, M <: AbstractNLPModel{T, S}, Meta <: AbstractNLPModelMeta{T, S}} <: AbstractPartiallySeparableNLPModel{T, S}

A partitioned quasi-Newton NLPModel. A PartiallySeparableNLPModel has field:

meta: gather information about the PartiallySeparableNLPModel;

part_data: allocate the partitioned structures required by a partitioned quasi-Newton trust-region method;

nlp: the original NLPModel;

counters: count how many standards methods of NLPModels are called.

NLPModels.grad! — Method

g = grad!(nlp, x, g)

Evaluate ∇f(x), the gradient of the objective function at x in place.

NLPModels.hprod! — Method

hprod!(nlp::PartiallySeparableNLPModel, x::AbstractVector, v::AbstractVector, Hv::AbstractVector; obj_weight=1.)

Evaluate the product of the objective Hessian at x with the vector v, with objective function scaled by obj_weight.

NLPModels.obj — Method

f = obj(nlp, x)

Evaluate f(x), the objective function of nlp at x.

PartiallySeparableNLPModels.Mod_partitionedNLPModel.get_expr_tree — Method

expr, n, x0 = get_expr_tree(adnlp::MathOptNLPModel; x0::Vector{T} = copy(adnlp.meta.x0), kwargs...) where {T <: Number}
expr, n, x0 = get_expr_tree(adnlp::ADNLPModel; x0::Vector{T} = copy(adnlp.meta.x0), kwargs...) where {T <: Number}

Return for a MathOptNLPModel or a ADNLPModel: the expression tree expr::Expr of the objective function, the size of the problem n and the initial point x0.

PartiallySeparableNLPModels.Mod_partitionedNLPModel.hess_approx — Method

B = hess_approx(nlp::PartiallySeparableNLPModel)

Return the Hessian approximation of nlp.

PartiallySeparableNLPModels.Mod_PQN.PartitionedDataTRPQN — Type

PartitionedDataTRPQN{G, T <: Number, P <: Part_mat{T}} <: PartitionedData

Gather the structures required to run a partitioned quasi-Newton trust-region method. PartitionedDataTRPQN has fields:

n: the size of the problem;

N: the number of element functions;

vec_elt_fun: a ElementFunction vector, of size N;

M: the number of distinct element-function expression trees;

vec_elt_complete_expr_tree: a Complete_expr_tree vector, of size M;

element_expr_tree_table: a vector of size M, the i-th element element_expr_tree_table[i]::Vector{Int} informs which element functions use the vec_elt_complete_expr_tree[i] expression tree;

index_element_tree: a vector of size N where each component indicates which Complete_expr_tree from vec_elt_complete_expr_tree is used for the corresponding element;

vec_compiled_element_gradients: the vector gathering the compiled tapes for every element gradient evaluations;

x: the current point;

v: a temporary vector;

s: the current step;

pg: the partitioned gradient;

pv: a temporary partitioned vector;

py: the partitioned gradient difference;

ps: the partitioned step;

phv: the partitioned Hessian-vector product;

pB: the partitioned matrix (main memory cost);

fx: the current value of the objective function;

name: the name of partitioned quasi-Newton update performed at each iteration:

name=:pbfgs: every element-Hessian approximation is updated with BFGS;

name=:psr1: every element-Hessian approximation is updated with SR1;

name=:pse: every element-Hessian approximation is updated with BFGS if the curvature condition holds, or with SR1 otherwise;

name=:pcs: each element-Hessian approximation with BFGS if it is classified as convex, or with SR1 otherwise;

name=:plbfgs: every element-Hessian approximations is a LBFGS operator;

name=:plsr1: every element-Hessian approximations is a LSR1 operator;

name=:plse: by default, every element-Hessian approximations is a LBFGS operator as long as the curvature condition holds, otherwise it becomes a LSR1 operator.

PartiallySeparableNLPModels.Mod_PQN.build_PartitionedDataTRPQN — Method

partitioneddata_tr_pqn = build_PartitionedDataTRPQN(expr_tree, n)

Return the structure required to run a partitioned quasi-Newton trust-region method. It finds the partially-separable structure of an expression tree expr_tree representing f(x) = ∑fᵢ(xᵢ). Then it allocates the partitioned structures required. To define properly the sparse matrix of the partitioned matrix we need the size of the problem: n.

PartiallySeparableNLPModels.Mod_PQN.update_PQN! — Method

update_PQN!(pd_pqn::PartitionedDataTRPQN{G, T, P}, s::Vector{T})
update_PQN!(pd_pqn::PartitionedDataTRPQN{G, T, P}, x::Vector{T}, s::Vector{T})

Perform the partitioned quasi-Newton update given the current point x and the step s. When x is omitted, update_PQN! consider that pd_pqn has the current point in pdpqn.x. Moreover, it assumes that the partitioned gradient atxis already computed inpdpqn.pg`.

PartiallySeparableNLPModels.Mod_PQN.update_PQN — Method

B = update_PQN(pd_pqn::PartitionedDataTRPQN{G, T, P}, x::Vector{T}, s::Vector{T};

Perform the partitioned quasi-Newton update given the current point x and the step s.

PartiallySeparableNLPModels.Mod_ab_partitioned_data.update_nlp! — Method

update_nlp!(pd_pqn::PartitionedDataTRPQN{G, T, P}, s::Vector{T})
update_nlp!(pd_pqn::PartitionedDataTRPQN{G, T, P}, x::Vector{T}, s::Vector{T})

Perform the partitioned quasi-Newton update given the current point x and the step s. When x is omitted, update_PQN! consider that pd_pqn has the current point in pdpqn.x. Moreover, it assumes that the partitioned gradient atxis already computed inpdpqn.pg`.