Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

PartiallySeparableNLPModels.Hv

PartiallySeparableNLPModels.Hv2

PartiallySeparableNLPModels.Hv_only_product

PartiallySeparableNLPModels.Mod_PBFGS.build_PartitionedData_TR_PBFGS

PartiallySeparableNLPModels.Mod_PBFGS.update_PBFGS

PartiallySeparableNLPModels.Mod_PBFGS.update_PBFGS!

PartiallySeparableNLPModels.Mod_PLBFGS.build_PartitionedData_TR_PLBFGS

PartiallySeparableNLPModels.Mod_PLBFGS.update_PLBFGS

PartiallySeparableNLPModels.Mod_PLBFGS.update_PLBFGS!

PartiallySeparableNLPModels.Mod_ab_partitioned_data.evaluate_grad_part_data

PartiallySeparableNLPModels.Mod_ab_partitioned_data.evaluate_y_part_data!

PartiallySeparableNLPModels.Mod_ab_partitioned_data.product_part_data_x

PartiallySeparableNLPModels.Mod_common.compiled_grad_elmt_fun

PartiallySeparableNLPModels.Mod_common.distinct_element_expr_tree

PartiallySeparableNLPModels.build_gradient

PartiallySeparableNLPModels.check_Inf_Nan

PartiallySeparableNLPModels.check_Inf_Nan

PartiallySeparableNLPModels.compiled_grad_of_elmt_fun

PartiallySeparableNLPModels.construct_Sparse_Hessian

PartiallySeparableNLPModels.construct_set_index_element_view!

PartiallySeparableNLPModels.construct_views

PartiallySeparableNLPModels.create_pre_compiled_tree

PartiallySeparableNLPModels.deduct_partially_separable_structure

PartiallySeparableNLPModels.element_gradient!

PartiallySeparableNLPModels.element_gradient2!

PartiallySeparableNLPModels.eval_ni_ones

PartiallySeparableNLPModels.evaluate_SPS

PartiallySeparableNLPModels.evaluate_SPS_gradient

PartiallySeparableNLPModels.evaluate_SPS_gradient!

PartiallySeparableNLPModels.evaluate_SPS_gradient2!

PartiallySeparableNLPModels.evaluate_element_hessian

PartiallySeparableNLPModels.evaluate_gradient

PartiallySeparableNLPModels.evaluate_hessian

PartiallySeparableNLPModels.evaluate_obj_pre_compiled

PartiallySeparableNLPModels.get_elmnt_fun_index_view

PartiallySeparableNLPModels.get_index_deleted

PartiallySeparableNLPModels.get_related_function

PartiallySeparableNLPModels.get_related_var

PartiallySeparableNLPModels.grad_ni_to_n!

PartiallySeparableNLPModels.id_hessian!

PartiallySeparableNLPModels.minus_grad_vec!

PartiallySeparableNLPModels.product_matrix_sps

PartiallySeparableNLPModels.product_matrix_sps!

PartiallySeparableNLPModels.product_vector_sps

PartiallySeparableNLPModels.struct_hessian

PartiallySeparableNLPModels.struct_hessian!

​

PartiallySeparableNLPModels.Hv — Method

Hv(sps, x, v)

Compute the product hessian vector of the hessian at the point x dot the vector v. This version both ReverseDiff and ForwardDiff.

PartiallySeparableNLPModels.Hv2 — Method

Hv2(sps, x, v)

Compute the product hessian vector of the hessian at the point x dot the vector v. This version uses only ReverseDiff.

PartiallySeparableNLPModels.Hv_only_product — Method

Hv_only_product(sps, Hess_matrix, x)

return the product dot(Hessmatrix, x), the function fits with the partitionned representation of the matrix Hessmatrix. In order to do the right operations in the right order we need the partially separable structure sps. Less efficient than productmatrixsps.

PartiallySeparableNLPModels.build_gradient — Method

build_gradient(sps, g)

Constructs a vector of size n from the list of element gradient of the sps structure which has numerous element gradient of size nᵢ. The purpose of the function is to gather these element gradient into a real gradient of size n. The function gradnito_n! will add element gradient of size nᵢ at the right inside the gradient of size n.

PartiallySeparableNLPModels.check_Inf_Nan — Method

check_Inf_Nan(B)

function that check if an Hess_matrix is not full of Nan.

PartiallySeparableNLPModels.check_Inf_Nan — Method

check_Inf_Nan(B)

function that check if an array contains a Nan.

PartiallySeparableNLPModels.compiled_grad_of_elmt_fun — Method

compiled_grad_of_elmt_fun(elmt_fun)

Return the GradientTape compiled to speed up the ReverseDiff computation of the elmt_fun gradient in the future

PartiallySeparableNLPModels.construct_Sparse_Hessian — Method

construct_Sparse_Hessian(sps, B)

Build from the Partially separable Structure sps and the Hessian approximation B a SpaseArray which represent B in other form.

PartiallySeparableNLPModels.construct_set_index_element_view! — Method

construct_set_index_element_view(elmt_fun_vector, diff_elmt_tree)

For each elmtfun ∈ elmtfunvector we set the attribute indexelementtreeposition. For doing this we use the function getelmntfunindexview(elmtfunvector, diffelmttree) that builds a vector of size length(elmtfunvector) with the right indexes.

PartiallySeparableNLPModels.construct_views — Method

construct_views(x, related_vars)

related_vars a vector (from different element tree) of vector (several element function linked with each tree)

of vector{int} (a set of index, which represent the elemental variable).

x a point ∈ Rⁿ

returns a view of x for each set of index representing the elemental variable.

PartiallySeparableNLPModels.create_pre_compiled_tree — Method

create_pre_compiled_tree(vector_calculus_trees, vector_x_views)

create a precompiled tree for a particuliar number of evaluation. The size of vectorcalculustrees and vectorxviews must match. Each calculustree from vectorcalculustrees is pre-compiled for nᵢ evaluation, nᵢ the size of the corresponding element of vectorx_views

PartiallySeparableNLPModels.deduct_partially_separable_structure — Function

deduct_partially_separable_structure(expr_tree, n)

Find the partially separable structure of a function f stored as an expression tree expr_tree. To define properly the size of sparse matrix we need the size of the problem : n. At the end, we get the partially separable structure of f, f(x) = ∑fᵢ(xᵢ)

PartiallySeparableNLPModels.element_gradient! — Method

element_gradient!(compil_tape, x, g)

Compute the element grandient from the compil_tape compiled before according to the vector x, and store the result in the vector g Use of ReverseDiff

PartiallySeparableNLPModels.element_gradient2! — Method

element_gradient2!(expr_tree, x, g)

Compute the element grandient of the function represents by expr_tree according to the vector x, and store the result in the vector g. This was the previous version using ForwardDiff. The actual version using ReverseDiff is more efficient.

PartiallySeparableNLPModels.eval_ni_ones — Method

get_different_CalculusTree( all_element_tree, element_vars)

Selects the distinct tree from the allelementtree to create a Vector of distinct element tree: differentcalculustree. In addition to that create a Vector of index to linked the tree from allelementtree to differentcalculustree : differentcalculustreeindex. elementvars is used to count the numbers of variables used in a tree, if the number of variables is different we don't test the tree and avoid to test the equality between trees. length(differentcalculustreeindex) == length(allelementtree) == length(elementvars) return (differentcalculustree, differentcalculustree_index)

PartiallySeparableNLPModels.evaluate_SPS — Method

evaluate_SPS(sps,x)

Evaluate the structure sps on the point x ∈ Rⁿ. Since we work on subarray of x, we allocate x to the structure sps in a first time. Once this step is done we select the calculus tree needed as well as the view of x needed. Then we evaluate the different calculus tree on the needed point (since the same function appear a lot of time). At the end we sum the result

PartiallySeparableNLPModels.evaluate_SPS_gradient! — Method

evaluate_SPS_gradient!(sps,x,g)

Compute the gradient of the partially separable structure sps, and store the result in the gradvector structure g. Using ReversDiff package. Not obvious good behaviour with Threads.@threads, sometime yes sometime no. Noted that we use the previously compiled GradientTape in elementgradient! that use ReverseDiff.

PartiallySeparableNLPModels.evaluate_SPS_gradient — Method

evaluate_SPS_gradient(sps,x)

Return the gradient of the partially separable structure sps at the point x. Using ReversDiff package. Not obvious good behaviour with Threads.@threads, sometime yes sometime no. Noted that we use the previously compiled GradientTape in element_gradient! that use ReverseDiff.

PartiallySeparableNLPModels.evaluate_SPS_gradient2! — Method

evaluate_SPS_gradient2!(sps,x,g)

Compute the gradient of the partially separable structure sps, and store the result in the grad_vector structure g. Using ForwardDiff package. Bad behaviour with Threads.@threads. This was the previous version using ForwardDiff. The actual version using ReverseDiff is more efficient.

PartiallySeparableNLPModels.evaluate_element_hessian — Method

evaluate_element_hessian(fᵢ,xᵢ)

Compute the Hessian of the elemental function fᵢ : Gᵢ a nᵢ × nᵢ matrix. So xᵢ a vector of size nᵢ. The result of the function is the triplet of the sparse matrix Gᵢ.

PartiallySeparableNLPModels.evaluate_gradient — Method

evaluate_gradient(sps,x)

evalutate the gradient of the partially separable function f = ∑ fι, stored in the sps structure at the point x, return a vector of size n (the number of variable) which is the gradient. Première version de la fonction inutile car inefficace.

PartiallySeparableNLPModels.evaluate_hessian — Method

evaluate_hessian(sps,x)

evalutate the hessian of the partially separable function f = ∑ fᵢ, stored in the sps structure at the point x. Return the sparse matrix of the hessian of size n × n.

PartiallySeparableNLPModels.evaluate_obj_pre_compiled — Method

evaluate_obj_pre_compiled(sps, point_x)

Evaluate the structure sps at pointx using the precompiled element tree for nᵢ evaluation. This function is the current evaluateSPS.

PartiallySeparableNLPModels.get_elmnt_fun_index_view — Method

get_elmnt_fun_index_view(vector_element_fun, different_element_tree)

For each elementfun ∈ vectorelementfun, returns the position of the linked element tree corresponding in differentelement_tree. It will be use later for retrieve a view of the variable used by each element function.

PartiallySeparableNLPModels.get_index_deleted — Method

get_index_deleted(tree_vector, tree, numbervar, vector_numbervar)

retrieve the indexes of treevector such that for all i ∈ indexes treevector[i] == tree. This function used numbervar and vector_numbervar to speed up the comparison by avoiding equality between trees.

PartiallySeparableNLPModels.get_related_function — Method

get_related_function(elmt_functionS, calculusTreeS)

Cette fonction récupère les indices des fonctions éléments elmt_functionS lié aux quelques arbres de calculs calculusTreeS de la structure

PartiallySeparableNLPModels.get_related_var — Method

get_related_var(vec_elmt_fun, index_elmt_tree)

Renvoie un tableau d'entier de 3 dimensions. La première dimension de taille length(indexelmttree) correspond au nombre de fonctions éléments distinctes (arbre de calcul). Pour chaque fonction éléments distinctes e nous avons un vecteur vₑ :: Vector{Vector{Int}} tel que length(vₑ) == length(indexelmttree[e]) indexelmttree[e] contient les indices des fonctions éléments rattachés à la fonction élément e. Le résultat de la fonction renverra pour chaque indice de fonctions éléments les variables utilisées ( :: Vector{Int}) par celle-ci ce qui nous donne un tableau 3 dimensions.

PartiallySeparableNLPModels.grad_ni_to_n! — Method

grad_ni_to_n!(element_gradient, used_var, gradient)

Add to the gradient the value of elementgradient according to the vector of usedvar given. length(usedvar) == length(elementgradient)

PartiallySeparableNLPModels.id_hessian! — Method

id_hessian!(sps, B)

Construct a kinf of Id Hessian, it will initialize each element Hessian Bᵢ with an Id matrix, B = ∑ᵢᵐ Uᵢᵀ Bᵢ Uᵢ

PartiallySeparableNLPModels.minus_grad_vec! — Method

minus_grad_vec!(g1,g2,res)

Store in res: g1 minus g2, but g1 and g2 have a particular structure which is grad_vector{T}. We need this operation to have the difference for each element gradient for TR method. g1 = gₖ and g2 = gₖ₋₁.

PartiallySeparableNLPModels.product_matrix_sps! — Method

product_matrix_sps!(sps,B,x,Bx)

This function make the product of the structure B which represents a symetric matrix and the vector x, the result is stored in Bx. We need the structure sps for the variable used in each B[i], to replace B[i]*x[i] in the result vector by using f_inter!.

PartiallySeparableNLPModels.product_matrix_sps — Method

product_matrix_sps(sps,B,x)

This function make the product of the structure B which represents a symetric matrix and the vector x. We need the structure sps for the variable used in each B[i], to replace B[i]*x[i] in the result vector.

PartiallySeparableNLPModels.product_vector_sps — Method

product_vector_sps(sps, g, x)

compute the product g⊤ x = ∑ Uᵢ⊤ gᵢ⊤ xᵢ. So we need the sps structure to get the Uᵢ. On ne s'en sert pas en pratique mais peut-être pratique pour faire des vérifications

PartiallySeparableNLPModels.struct_hessian! — Method

struct_hessian!(sps,x,H)

Evalutate the hessian of the partially separable function, stored in the sps structure at the point x. Store the Hessian in a particular structure H :: Hess_matrix.

PartiallySeparableNLPModels.struct_hessian — Method

struct_hessian(sps,x)

evalutate the hessian of the partially separable function, stored in the sps structure at the point x. Return the Hessian in a particular structure : Hess_matrix.

PartiallySeparableNLPModels.Mod_ab_partitioned_data.evaluate_grad_part_data — Method

evaluate_grad_part_data(part_data,x)

Build the gradient vector at the point x from the element gradient computed and stored in part_data.pg .

PartiallySeparableNLPModels.Mod_ab_partitioned_data.evaluate_y_part_data! — Method

evaluate_y_part_data!(part_data,x,s)
	evaluate_y_part_data!(part_data,s)

Compute the element gradients differences such as ∇̂fᵢ(x+s)-∇̂fᵢ(x) for each element functions. It stores the results in partdata.pv. evaluateypartdata!(part_data,s) consider that pg is alreagy the gradient of the point x

PartiallySeparableNLPModels.Mod_ab_partitioned_data.product_part_data_x — Method

product_part_data_x!(part_data, x)

Return the product of the partitioned matrix part_data*x.

PartiallySeparableNLPModels.Mod_PBFGS.build_PartitionedData_TR_PBFGS — Method

build_PartitionedData_TR_PBFGS(expr_tree, n)

Find the partially separable structure of a function f stored as an expression tree expr_tree. To define properly the size of sparse matrix we need the size of the problem : n. At the end, we get the partially separable structure of f, f(x) = ∑fᵢ(xᵢ)

PartiallySeparableNLPModels.Mod_PBFGS.update_PBFGS! — Method

update_PBFGS(pd_pbfgs,s)

Perform the PBFGS update givent the current iterate x and the next iterate s

PartiallySeparableNLPModels.Mod_PBFGS.update_PBFGS — Method

update_PBFGS(pd_pbfgs,x,s)

Perform the PBFGS update givent the two iterate x and s

PartiallySeparableNLPModels.Mod_PLBFGS.build_PartitionedData_TR_PLBFGS — Method

build_PartitionedData_TR_PLBFGS(expr_tree, n)

Find the partially separable structure of a function f stored as an expression tree expr_tree. To define properly the size of sparse matrix we need the size of the problem : n. At the end, we get the partially separable structure of f, f(x) = ∑fᵢ(xᵢ)

PartiallySeparableNLPModels.Mod_PLBFGS.update_PLBFGS! — Method

update_PLBFGS(pd_pblfgs,s)

Perform the PBFGS update givent the current iterate x and the next iterate s

PartiallySeparableNLPModels.Mod_PLBFGS.update_PLBFGS — Method

update_PLBFGS(pd_pblfgs,x,s)

Perform the PBFGS update givent the two iterate x and s

PartiallySeparableNLPModels.Mod_common.compiled_grad_elmt_fun — Method

compiledgradelmtfun(elmtfun, ni) Return the GradientTape compiled to speed up the ReverseDiff computation of the elmt_fun gradient in the future

PartiallySeparableNLPModels.Mod_common.distinct_element_expr_tree — Method

distinct_element_expr_tree(vec_element_expr_tree, vec_element_variables; N)

Filter the vector vecelementexprtree to obtain only the element functions that are distincts as elementexprtree. length(elementexprtree) == M. In addition it returns indexelement_tree, who records the index 1 <= i <= M of each element function